"let" added to for loops! Asteroid splits.
Had a bug in Asteroids.breakUp until I added "let" 
to the variable declaration in the for loop. 
Apparently the scope of the i was much broader 
than I realized, and it broke a larger loop.

Now the asteroid splits into 2 after being shot, and 
it has a minimum size.  These parameters can be 
configured in:
Asteroid.numChildren = 2
Asteroid.minParentSize = 10
Asteroid.childSizeRatio = 0.5